refactor: keep makeNativeCameraStop within Camera
